pyproject.toml
[tool.poetry]
name = "llm-rpg"
version = "0.1.0"
description = ""
authors = ["WoutVossen <vossen.w@hotmail.com>"]
readme = "README.md"
packages = [{include = "llm_rpg", from = "src"}]
include = [
    { path = "src/llm_rpg/assets/**/*", format = "wheel" },
    { path = "src/llm_rpg/assets/**/*", format = "sdist" },
]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
openai = "^1.57.1"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.2"
ollama = "^0.5.1"
pygame = "^2.6.1"
unfake = "^1.0.7"
torchvision = "^0.24.1"
diffusers = "^0.36.0"
transformers = ">=4.57.3"
tokenizers = ">=0.22.1"
peft = "^0.18.0"
playwright = "^1.48.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pre-commit = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
